Moved existing mail logic to mailerService
