+++ src/Intracto/SecretSantaBundle/Entity/EntryService.php
@@ -78,67 +78,6 @@ class EntryService
     }
 
     /**
-     * Sends out all mails for a Pool
-     *
-     * @param Pool $pool
-     */
-    public function sendSecretSantaMailsForPool(Pool $pool)
-    {
-        $pool->setSentdate(new \DateTime('now'));
-        $this->em->flush($pool);
-
-        foreach ($pool->getEntries() as $entry) {
-            $this->sendSecretSantaMailForEntry($entry);
-        }
-    }
-
-    /**
-     * Sends out mail for a Entry
-     *
-     * @param Entry $entry
-     */
-    public function sendSecretSantaMailForEntry(Entry $entry)
-    {
-        $this->translator->setLocale($entry->getPool()->getLocale());
-
-        $message = $entry->getPool()->getMessage();
-        $message = str_replace('(NAME)', $entry->getName(), $message);
-        $message = str_replace('(ADMINISTRATOR)', $entry->getPool()->getOwnerName(), $message);
-        $txtBody = $this->templating->render(
-            'IntractoSecretSantaBundle:Emails:secretsanta.txt.twig',
-            array('message' => $message, 'entry' => $entry)
-        );
-        $htmlBody = $this->templating->render(
-            'IntractoSecretSantaBundle:Emails:secretsanta.html.twig',
-            array('message' => $message, 'entry' => $entry)
-        );
-
-        $mail = \Swift_Message::newInstance()
-            ->setSubject($this->translator->trans('emails.secretsanta.subject'))
-            ->setFrom($this->adminEmail, $entry->getPool()->getOwnerName())
-            ->setReplyTo([$entry->getPool()->getOwnerEmail() => $entry->getPool()->getOwnerName()])
-            ->setTo($entry->getEmail(), $entry->getName())
-            ->setBody($txtBody)
-            ->addPart($htmlBody, 'text/html');
-        $this->mailer->send($mail);
-    }
-
-    /**
-     * @param Pool $pool
-     */
-    public function sendPoolMatchesToAdmin(Pool $pool)
-    {
-        $this->translator->setLocale($pool->getLocale());
-        $this->mailer->send(\Swift_Message::newInstance()
-            ->setSubject($this->translator->trans('emails.admin_matches.subject'))
-            ->setFrom($this->adminEmail, $this->translator->trans('emails.sender'))
-            ->setTo($pool->getOwnerEmail(), $pool->getOwnerName())
-            ->setBody($this->templating->render('IntractoSecretSantaBundle:Emails:admin_matches.html.twig', array('pool' => $pool)), 'text/html')
-            ->addPart($this->templating->render('IntractoSecretSantaBundle:Emails:admin_matches.txt.twig', array('pool' => $pool)), 'text/plain')
-        );
-    }
-
-    /**
      * @return array
      *
      * @throws \Doctrine\DBAL\DBALException
