+++ src/Intracto/SecretSantaBundle/Resources/config/services.xml
@@ -40,5 +40,13 @@
             <argument type="service" id="intracto_secret_santa.entry"></argument>
             <argument type="service" id="intracto_secret_santa.wishlist"></argument>
         </service>
+        <service id="intracto_secret_santa.mail" class="Intracto\SecretSantaBundle\Mailer\MailerService">
+            <argument type="service" id="mailer"></argument>
+            <argument type="service" id="doctrine.orm.entity_manager"></argument>
+            <argument type="service" id="templating"></argument>
+            <argument type="service" id="translator"></argument>
+            <argument type="service" id="router"></argument>
+            <argument>%admin_email%</argument>
+        </service>
     </services>
 </container>
